Windows Update Reset Version 3.1.5 Beta 1 Test
Changes:
Added a file organizer.
Added more warnings when destroying windows.

Notes:
This is a test beta! Do not expect this to be stable.
This is only the batch file. Most of the stuff like the readme are not included.
This version will not count as a new version in Version 3.1.4.